refactor(c): convert setup.cfg to pyproject.toml
Drop setup.py: move the build configuration to a build backend module.
