+++ psycopg_c/pyproject.toml
@@ -1,6 +1,12 @@
 [build-system]
 requires = [
-    "setuptools >= 49.2.0",
+    # Note: pinning these versions strictly because of the setuptools warning:
+    #
+    #   `[tool.setuptools.ext-modules]` in `pyproject.toml` is still
+    #   *experimental* and likely to change in future releases
+    #
+    "setuptools == 75.6.0; python_version >= '3.9'",
+    "setuptools == 75.3.0; python_version < '3.9'",  # last supported version
     "wheel >= 0.37",
     "tomli >= 2.0.1; python_version < '3.11'",
 ]
@@ -15,3 +21,86 @@ backend-path = ["build_backend"]
 [cython-backend]
 # These packages are only installed if there are pyx files to compile.
 cython-requires = ["Cython >= 3.0.0"]
+
+[project]
+name = "psycopg-c"
+description = "PostgreSQL database adapter for Python -- C optimisation distribution"
+version = "3.2.4.dev1"
+classifiers = [
+    "Development Status :: 5 - Production/Stable",
+    "Intended Audience :: Developers",
+    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
+    "Operating System :: MacOS :: MacOS X",
+    "Operating System :: Microsoft :: Windows",
+    "Operating System :: POSIX",
+    "Programming Language :: Cython",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: Implementation :: CPython",
+    "Topic :: Database",
+    "Topic :: Database :: Front-Ends",
+    "Topic :: Software Development",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+]
+requires-python = ">= 3.8"
+
+[[project.authors]]
+name = "Daniele Varrazzo"
+email = "daniele.varrazzo@gmail.com"
+
+[project.license]
+text = "GNU Lesser General Public License v3 (LGPLv3)"
+
+[project.urls]
+Homepage = "https://psycopg.org/"
+Documentation = "https://psycopg.org/psycopg3/docs/"
+Changes = "https://psycopg.org/psycopg3/docs/news.html"
+Code = "https://github.com/psycopg/psycopg"
+"Issue Tracker" = "https://github.com/psycopg/psycopg/issues"
+Download = "https://pypi.org/project/psycopg-c/"
+
+[project.readme]
+file = "README.rst"
+content-type = "text/x-rst"
+
+[tool.setuptools]
+packages = [
+    "psycopg_c",
+    "psycopg_c.pq",
+    "psycopg_c._psycopg",
+    "psycopg_c.types",
+]
+zip-safe = false
+license-files = ["LICENSE.txt"]
+include-package-data = true
+
+[tool.setuptools.package-data]
+psycopg_c = [
+    "py.typed",
+    "*.pyi",
+    "*.pxd",
+    "_psycopg/*.pxd",
+    "pq/*.pxd",
+]
+psycopg_binary = [
+    "py.typed",
+    "*.pyi",
+]
+
+# Note: these ext modules lack details such as libraries and directories.
+# They are added by the psycopg_build_ext build module.
+[[tool.setuptools.ext-modules]]
+name = "psycopg_c._psycopg"
+sources = ["psycopg_c/_psycopg.c", "psycopg_c/types/numutils.c"]
+
+[[tool.setuptools.ext-modules]]
+name = "psycopg_c.pq"
+sources = ["psycopg_c/pq.c"]
+
+[tool.setuptools.cmdclass]
+build_ext = "psycopg_build_ext.psycopg_build_ext"
