--- psycopg_c/setup.cfg
@@ -1,62 +1,0 @@
-[metadata]
-name = psycopg-c
-description = PostgreSQL database adapter for Python -- C optimisation distribution
-url = https://psycopg.org/psycopg3/
-author = Daniele Varrazzo
-author_email = daniele.varrazzo@gmail.com
-license = GNU Lesser General Public License v3 (LGPLv3)
-version = 3.2.4.dev1
-
-project_urls =
-    Homepage = https://psycopg.org/
-    Documentation = https://psycopg.org/psycopg3/docs/
-    Changes = https://psycopg.org/psycopg3/docs/news.html
-    Code = https://github.com/psycopg/psycopg
-    Issue Tracker = https://github.com/psycopg/psycopg/issues
-    Download = https://pypi.org/project/psycopg-c/
-
-classifiers =
-    Development Status :: 5 - Production/Stable
-    Intended Audience :: Developers
-    License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)
-    Operating System :: MacOS :: MacOS X
-    Operating System :: Microsoft :: Windows
-    Operating System :: POSIX
-    Programming Language :: Cython
-    Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
-    Programming Language :: Python :: 3.11
-    Programming Language :: Python :: 3.12
-    Programming Language :: Python :: 3.13
-    Programming Language :: Python :: Implementation :: CPython
-    Topic :: Database
-    Topic :: Database :: Front-Ends
-    Topic :: Software Development
-    Topic :: Software Development :: Libraries :: Python Modules
-
-long_description = file: README.rst
-long_description_content_type = text/x-rst
-license_files = LICENSE.txt
-
-[options]
-python_requires = >= 3.8
-packages = find:
-zip_safe = False
-
-[options.package_data]
-# NOTE: do not include .pyx files: they shouldn't be in the sdist
-# package, so that build is only performed from the .c files (which are
-# distributed instead).
-psycopg_c =
-    py.typed
-    *.pyi
-    *.pxd
-    _psycopg/*.pxd
-    pq/*.pxd
-
-# In the psycopg-binary distribution don't include cython-related files.
-psycopg_binary =
-    py.typed
-    *.pyi
