+++ tests/constraints.txt
@@ -30,7 +30,6 @@ sphinx-autobuild == 2021.3.14
 sphinx-autodoc-typehints == 1.12.0
 
 # Build tools
-setuptools == 49.2.0
 wheel == 0.37
 Cython == 3.0.0
 tomli == 2.0.1
